+++ scru128.go
@@ -16,22 +16,23 @@
 // See SCRU128 Specification for details: https://github.com/scru128/spec
 package scru128
 
-// Maximum value of 48-bit timestamp field.
+// The maximum value of 48-bit timestamp field.
 const maxTimestamp uint64 = 0xffff_ffff_ffff
 
-// Maximum value of 24-bit counter_hi field.
+// The maximum value of 24-bit counter_hi field.
 const maxCounterHi uint32 = 0xff_ffff
 
-// Maximum value of 24-bit counter_lo field.
+// The maximum value of 24-bit counter_lo field.
 const maxCounterLo uint32 = 0xff_ffff
 
-var defaultGenerator = NewGenerator()
+var globalGenerator = NewGenerator()
 
-// Generates a new SCRU128 ID object, or panics if crypto/rand fails.
+// Generates a new SCRU128 ID object using the global generator, or panics if
+// crypto/rand fails.
 //
 // This function is thread-safe; multiple threads can call it concurrently.
 func New() Id {
-	id, err := defaultGenerator.Generate()
+	id, err := globalGenerator.Generate()
 	if err != nil {
 		panic(err)
 	}
@@ -39,7 +40,7 @@ func New() Id {
 }
 
 // Generates a new SCRU128 ID encoded in the 26-digit canonical string
-// representation, or panics if crypto/rand fails.
+// representation using the global generator, or panics if crypto/rand fails.
 //
 // This function is thread-safe. Use this to quickly get a new SCRU128 ID as a
 // string.